requirements.txt
alembic==1.2.1
aniso8601==8.0.0
Click==7.0
Flask==1.1.1
Flask-Migrate==2.4.0
Flask-Bcrypt==0.7.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.2.1
psycopg2-binary==2.7.7
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.3
PyJWT==1.7.1
six==1.12.0
SQLAlchemy==1.3.9
Werkzeug==0.16.0
selenium
gunicorn==19.9.0
